Intersection of Two Linked Lists
Intersection of Two Linked Lists

In a system there are two singly linked list.
By some programming error, the end node of one of the linked lists got linked to the second list, forming an inverted Y-shaped list. 
Write a program to get the point where two linked lists merge.
We have to find the intersection part in this system.